+++ src/components/Chat/ChatContent/Message/CommandPrompt/CommandPrompt.tsx
@@ -1,9 +1,11 @@
 import React, { useEffect, useRef, useState } from 'react';
 import useStore from '@store/store';
-import { hideOnClickOutside } from '@utils/handleClick';
+
 import { useTranslation } from 'react-i18next';
 import { matchSorter } from 'match-sorter';
 import { Prompt } from '@type/prompt';
+
+import useHideOnOutsideClick from '@hooks/useHideOnOutsideClick';
 
 const CommandPrompt = ({
   _setContent,
@@ -12,10 +14,10 @@ const CommandPrompt = ({
 }) => {
   const { t } = useTranslation();
   const prompts = useStore((state) => state.prompts);
-  const [dropDown, setDropDown] = useState<boolean>(false);
   const [_prompts, _setPrompts] = useState<Prompt[]>(prompts);
   const [input, setInput] = useState<string>('');
-  const dropDownRef = useRef<HTMLDivElement>(null);
+
+  const [dropDown, setDropDown, dropDownRef] = useHideOnOutsideClick();
 
   useEffect(() => {
     const filteredPrompts = matchSorter(useStore.getState().prompts, input, {
@@ -28,10 +30,6 @@ const CommandPrompt = ({
     _setPrompts(prompts);
     setInput('');
   }, [prompts]);
-
-  useEffect(() => {
-    hideOnClickOutside(dropDownRef, dropDown, setDropDown);
-  }, [dropDownRef, dropDown]);
 
   return (
     <div className='relative max-wd-sm' ref={dropDownRef}>
